Generic flush.Runtime + consolidate engine/lsm root files
flush.Runtime is now generic over the payload type — flush.New[*memTable]
and flush.Task[*memTable] replace the any-typed payload. The hot-path
type assertion task.Payload.(*memTable) goes away and the lsm worker
loop reads task.Payload as a typed *memTable directly. flush.ErrNilPayload
is dropped (Go's type system already forbids untyped nil for typed
pointers in tests; production code holds a real *memTable).

Root-file cleanup:
  iterator.go    -> folded into memtable.go
  compactdef.go  -> folded into compaction.go (tests merge too)
  errors.go      -> sentinels moved into lsm.go
  diagnostics.go -> Diagnostics types and LSM.Diagnostics() to lsm.go,
                    range-filter metric helpers + entryCount to
                    level_manager.go, tablesSnapshot to level_handler.go

Each remaining root .go has its 1:1 test pair (plus options.go for
config types and lsm_bench_test.go for the standalone benches).
